+++ Anthony_John_Order_Parts_CaseStudy/src/main/java/com/example/Anthony_John_Order_Parts_CaseStudy/entity/RudderEntity.java
@@ -20,6 +20,6 @@ public class RudderEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int rudderId; // PK for rudder
+    private long rudderId; // PK for rudder
 
 }
