+++ Anthony_John_Order_Parts_CaseStudy/src/main/java/com/example/Anthony_John_Order_Parts_CaseStudy/entity/WingEntity.java
@@ -18,7 +18,7 @@ public class WingEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int wingId; // PK for wing
+    private long wingId; // PK for wing
 
     private String lWing; // designates the left wing
     private String rWing; // designates the right wing
